requirements.txt
# Core dependencies
google-generativeai>=0.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Web scraping and parsing
lxml>=4.9.0
html5lib>=1.1
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
asyncio-throttle>=1.0.2

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Enhanced text processing
nltk>=3.8.0
textblob>=0.17.0

# Optional: Additional web scraping capabilities
selenium>=4.15.0
playwright>=1.40.0
# requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
urllib3>=1.26.0
asyncio>=3.4.3
